Update predict algorithm
The 'share' aspect of the algorithm was very small for almost all movies even on this subset of the data. Using a larger subset would reduce it even more. I updated the algorithm by dividing the update by the max 'share'. This increased all of the increments without any of them going over 1.